% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{aggregateTests}
\alias{aggregateTests}
\title{Aggregate the various tests}
\usage{
aggregateTests(er, niter = 20000, discordantFactor = 5)
}
\arguments{
\item{er}{An `enrich.results` object}

\item{niter}{The number of iterations for the non-parametric p-value estimation}

\item{discordantFactor}{The factor by which to increase the mean p quantile of
discordant enrichments}
}
\value{
A results dataframe
}
\description{
Significance is assessed with a permutation of the p percentiles.
}
