% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{areamir}
\alias{areamir}
\title{areamir}
\usage{
areamir(signal, sets, ...)
}
\arguments{
\item{signal}{A named logical vector indicating which features are differentially-expressed}

\item{sets}{A data.frame with at least the following columns: 'set', 'feature', 'score'.}
}
\value{
a data.frame.
}
\description{
analytic Rank-based Enrichment Analysis using a conversion of the scores as 
weights.
}
