% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichMiR.R
\name{availableTests}
\alias{availableTests}
\title{availableTests}
\usage{
availableTests(x = NULL, sets = NULL)
}
\arguments{
\item{x}{Optional `x` input for `testEnrichment`}

\item{sets}{Optional `sets` input for `testEnrichment`}
}
\value{
A character vector of available tests
}
\description{
Returns the available enrichment tests, for the given inputs (if specified)
}
\examples{
availableTests()
}
