% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{breakStrings}
\alias{breakStrings}
\title{breakStrings}
\usage{
breakStrings(x, minSizeForBreak = 20, lb = "\\n")
}
\arguments{
\item{x}{A character vector.}

\item{minSizeForBreak}{The minimum number of characters for a string to be 
broken onto two lines (default 20).}

\item{lb}{The linebreak character to use (default \\n).}
}
\description{
Breaks each (long) component of a character vector into two lines.
}
