% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{CDplot}
\alias{CDplot}
\title{CDplot}
\usage{
CDplot(
  ll,
  by = NULL,
  k = 3,
  breaks = NULL,
  sameFreq = FALSE,
  addN = FALSE,
  dig.lab = NULL,
  minN = 10,
  pvals = FALSE,
  ...
)
}
\arguments{
\item{ll}{A named list of numeric values, or a vector of values to be divided}

\item{by}{If `ll` is a list, `by` is ignored. If `ll` is a vector of values, 
`by` should be a numeric/logical/factor vector of the same length according 
to which the values should be divided.}

\item{k}{The number of divisions (ignored if `ll` is a list, or if `by` is
logical or a factor).}

\item{breaks}{The breaks to use (ignored if `ll` is a list). If NULL, will
be calculated.}

\item{sameFreq}{Logical; whether to calculate breaks so thate the groups have
the same frequency (rather than same width)}

\item{addN}{Logical; whether to add group sizes to their names}

\item{dig.lab}{Number of digits for automatically-generated breaks.}

\item{minN}{The minimum number of items per group (groups below this will be
merged if the breaks are numeric).}

\item{pvals}{Logical; whether to print the p-values of KS tests between sets}

\item{...}{Passed to `geom_line` (can for instance be used for `size`, etc.)}
}
\value{
A ggplot.
}
\description{
Creates a cumulative distribution plot, eventually handling division of 
values into bins according to another variable.
}
