% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{CDplot3}
\alias{CDplot3}
\title{CDplot3}
\usage{
CDplot3(
  dea,
  sets,
  setName,
  line.size = 1.2,
  point.size = 0.8,
  checkSynonyms = TRUE,
  addN = FALSE,
  pvals = FALSE,
  ...
)
}
\arguments{
\item{dea}{A named numeric vector containing the signal for each feature, or
a DEA results data.frame (such as produced by major DEA packages, i.e.
edgeR, DESeq2 or limma).}

\item{sets}{The target sets object containing site specific information with 
a "type" column indicating the Site Type.}

\item{setName}{The name of the set to plot in `sets`.}

\item{line.size}{Size of the line.}

\item{point.size}{Size of the points.}

\item{addN}{Logical; whether to add group sizes to their names}

\item{pvals}{Logical; whether to print the p-values of KS tests between sets}

\item{...}{Any other argument passed to \code{\link{CDplot}}.}
}
\value{
A ggplot.
}
\description{
A wrapper around \code{\link{CDplot}} to work directly with DEA results and
split by site type
}
