% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{CDplotWrapper}
\alias{CDplotWrapper}
\title{CDplotWrapper}
\usage{
CDplotWrapper(
  dea,
  sets,
  setName,
  k = 3,
  addN = FALSE,
  by = c("auto", "sites", "score", "best_stype", "type"),
  sameFreq = NULL,
  line.size = 1.2,
  point.size = 0.8,
  checkSynonyms = TRUE,
  pvals = FALSE,
  ...
)
}
\arguments{
\item{dea}{A named numeric vector containing the signal for each feature, or
a DEA results data.frame (such as produced by major DEA packages, i.e.
edgeR, DESeq2 or limma).}

\item{sets}{The target sets object.}

\item{setName}{The name of the set to plot in `sets`.}

\item{k}{The number of groups to form. Only applicable if by = "sites" or
"score".}

\item{addN}{Logical; whether to add group sizes to their names}

\item{by}{The variable by which to form the groups; either "sites", "score",
"best_stype" or "type". If unspecified, will determine the best depending
on the available annotation data.}

\item{sameFreq}{Logical; whether to divide so as to obtained groups with
similar frequencies (rather than groups of similar width)}

\item{line.size}{Size of the line.}

\item{point.size}{Size of the points.}

\item{pvals}{Logical; whether to print the p-values of KS tests between sets}

\item{...}{Any other argument passed to \code{\link{CDplot}}.}
}
\value{
A ggplot.
}
\description{
A wrapper around \code{\link{CDplot}} to work directly with DEA results and a
target sets as inputs.
}
