% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadFunctions.R
\name{.fetch_Mirfamilies}
\alias{.fetch_Mirfamilies}
\title{Download Targetscan miRNA families}
\usage{
.fetch_Mirfamilies(species = c("human", "mouse", "rat"))
}
\description{
Download Targetscan miRNA families
}
