% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadFunctions.R
\name{.getTargetscan_miRfamilies}
\alias{.getTargetscan_miRfamilies}
\title{Download Targetscan8 miRNA families}
\usage{
.getTargetscan_miRfamilies(
  species = c("human", "mouse", "rat", "fish", "worm", "fly")
)
}
\description{
Download Targetscan8 miRNA families
}
