% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadFunctions.R
\name{.getTargetScanPos}
\alias{.getTargetScanPos}
\title{Downlaod Targetscan Position Files}
\usage{
.getTargetScanPos(
  species = c("human", "mouse", "rat"),
  incl_nonconsites = TRUE
)
}
\description{
Downlaod Targetscan Position Files
}
