% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadFunctions.R
\name{.getTargetScanPred}
\alias{.getTargetScanPred}
\title{Downlaod Targetscan8 Transcript / Gene Predictions}
\usage{
.getTargetScanPred(
  species = c("human", "mouse", "rat", "fly"),
  type = c("conserved", "all"),
  keepMers = FALSE
)
}
\description{
Downlaod Targetscan8 Transcript / Gene Predictions
}
