% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadFunctions.R
\name{.getTargetScanPred2}
\alias{.getTargetScanPred2}
\title{Downlaod Targetscan8 Transcript / Gene Predictions for fish and worm}
\usage{
.getTargetScanPred2(
  species = c("fish", "worm"),
  type = c("conserved", "all"),
  keepMers = FALSE
)
}
\description{
Downlaod Targetscan8 Transcript / Gene Predictions for fish and worm
}
