% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadFunctions.R
\name{.getTargetScanSites}
\alias{.getTargetScanSites}
\title{Downlaod Targetscan8 Site Position Files}
\usage{
.getTargetScanSites(
  species = c("human", "mouse", "rat"),
  incl_nonconsites = TRUE
)
}
\description{
Downlaod Targetscan8 Site Position Files
}
