% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{EA}
\alias{EA}
\title{EA}
\usage{
EA(tested, significant, TS, minSize = 5, testOnlyAnnotated = FALSE)
}
\arguments{
\item{tested}{The set of tested features (e.g. genes)}

\item{significant}{The set of differentially-expressed features.}

\item{TS}{A data.frame of miRNA targets, with at least the following columns: `family`, `rep.miRNA`, `feature`, `sites`.}

\item{minSize}{The minimum number of targets for a miRNA family to be tested (default 5).}

\item{testOnlyAnnotated}{Whether to excluded features that are bound by no miRNA (default FALSE).}
}
\value{
a data.frame.
}
\description{
Traditional overlap (Fisher test) between differentially-expressed features and miRNA targets.
}
