% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{ebayes}
\alias{ebayes}
\title{ebayes}
\usage{
ebayes(signal, sets, use.intercept = FALSE)
}
\arguments{
\item{signal}{A named numeric vector indicating the signal (e.g. logFC) of
each feature.}

\item{sets}{A data.frame of annotation, with at least the following columns: 
`set`, `feature` and `score`. Alternatively, a sparse numeric matrix, with sets as 
columns and features as rows (dimensions must be named).}

\item{use.intercept}{Logical; whether to use an intercept in the model.}
}
\value{
A data.frame.
}
\description{
A wrapper around \code{\link[limma]{lmFit}} and \code{\link[limma]{eBayes}}
for fitting the feature scores of many sets against a signal.
}
