% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{class}
\name{enrich.results}
\alias{enrich.results}
\alias{enrich.results-methods}
\alias{enrich.results-class}
\alias{initialize,enrich.results-method}
\alias{as.data.frame,enrich.results-method}
\alias{show,enrich.results-method}
\alias{summary,enrich.results-method}
\alias{$,enrich.results-method}
\alias{names,enrich.results-method}
\title{EnrichMiR enrichment results}
\usage{
\S4method{initialize}{enrich.results}(.Object, ...)

\S4method{as.data.frame}{enrich.results}(x, row.names = NULL, optional = FALSE, ...)

\S4method{show}{enrich.results}(object)

\S4method{summary}{enrich.results}(object)

\S4method{$}{enrich.results}(x, name)

\S4method{names}{enrich.results}(x)
}
\arguments{
\item{object, x}{An object of class \code{\link{enrich.results}}}
}
\value{
Depends on the method.
}
\description{
EnrichMiR enrichment results
}
\seealso{
\link{testEnrichment}, which is the normal way to create an 
  `enrich.results` object
}
