% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichMiR.R
\name{enrichMiR}
\alias{enrichMiR}
\title{enrichMiR}
\usage{
enrichMiR(
  DEA,
  TS,
  miRNA.expression = NULL,
  families = NULL,
  cleanNames = FALSE,
  ...
)
}
\arguments{
\item{DEA}{A data.frame of the results of a differential expression analysis, with features
(e.g. genes) as row names and with at least the following columns: `logFC`, `FDR`}

\item{TS}{A data.frame of miRNA targets, with at least the following columns: 
`family`, `rep.miRNA`, `feature`, `sites`. Alternatiely, the output of an 
aggregated scan of miRNA kdModels, with columns `transcript`, `seed`, 
`score` (or `log_kd`), `n.8mer`, etc.}

\item{miRNA.expression}{A named vector of miRNAs expression values. 
miRNAs not in this vector are assumed to be not expressed in the system, and are not tested.}

\item{families}{A named vector of miRNA families, with individual miRNAs as names. 
If not given, internal data from the package will be used (mouse miRNA families from targetScan).}

\item{cleanNames}{Logical; whether to remove prefix from all miRNA names (default FALSE).}

\item{...}{Passed to `testEnrichment`}
}
\value{
an enrich.results object.
}
\description{
A miRNA wrapper around `testEnrichment`, for continuity with previous 
enrichMiR versions.
}
