% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichMir.app.server.R
\name{enrichMiR.server}
\alias{enrichMiR.server}
\title{enrichMiR.server}
\usage{
enrichMiR.server(bData = NULL, logCallsFile = NULL)
}
\arguments{
\item{bData}{A named, nested list, the first level being species, the second
different binding annotations (see vignette for format details)}

\item{logCallsFile}{Optional path to a file when the date of enrichment calls
will be logged.}
}
\value{
A shiny server function
}
\description{
enrichMiR.server
}
