% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sublists.R
\name{enrichMiR2}
\alias{enrichMiR2}
\title{enrichMiR2}
\usage{
enrichMiR2(genes_in_set, allgenes, TS, categories,
  miRNA.expression = NULL, families = NULL,
  testOnlyAnnotated = FALSE, test = "siteMir", cleanNames = FALSE,
  minSize = 5)
}
\arguments{
\item{genes_in_set}{A vector of genes in the set of interest}

\item{allgenes}{A vector of all genes}

\item{TS}{A data.frame of miRNA targets, with at least the following columns: `family`, `rep.miRNA`, `feature`, `sites`.}

\item{categories}{A list of gene lists (e.g. GO categories), with names containing the set's label}

\item{miRNA.expression}{A named vector of miRNAs expression values. miRNAs not in this vector are assumed to be not expressed in the system, and are not tested.}

\item{families}{A named vector of miRNA families, with individual miRNAs as names. If not given, internal data from the package will be used (mouse miRNA families from targetScan).}

\item{testOnlyAnnotated}{Whether to excluded features that are bound by no miRNA (default FALSE).}

\item{test}{The test to be used, either `EA`, `wEA` and `siteMir` (default `siteMir`)}

\item{cleanNames}{Logical; whether to remove prefix from all miRNA names (default FALSE).}

\item{minSize}{The minimum effective size of a category for it to be tested (default 5)}
}
\value{
a data.frame
}
\description{
Looks for miRNA target enrichment across different subsets of genes
}
