% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalization.R
\name{findCoLocsubset}
\alias{findCoLocsubset}
\title{Find miRNA colocalizations in a subset of genes}
\usage{
findCoLocsubset(
  genes_in_set = NULL,
  expressed.genes = NULL,
  search.mode = c("MM", "MR", "MC"),
  cust.input = NULL,
  species = c("human", "mouse", "rat"),
  expressed.miRNAs = NULL,
  co_dist_min = NULL,
  co_dist_max = NULL,
  KD_Sites = FALSE,
  min_pairs_sub = 10
)
}
\arguments{
\item{genes_in_set}{A vector of genes, that are a subset of the expressed
genes. If specified, the colocalization function will test for an
enrichment of colocalizations in these "genes_in_set" with regards to the
gene universe ("expressed.genes").}

\item{expressed.genes}{A vector of expressed genes (at the moment EnsemblID,
maybe give option?) as a universe to search in}

\item{search.mode}{character object. Please specify here whether you want to
search for colocalizations between miRNAs & miRNAs (="MM") or between
miRNAs & RBPs (="MR") or between miRNAs & a custom position objects
(="MC"). See in cust.input which information this object has to contain.}

\item{cust.input}{A dataframe containing positional information of any given
object. The dataframe has to have at least 4 columns with the following
information and structure: Ensembl Transcript IDs in the first column (eg.
ENSMUST00000046533), the Motif or Seed specifier of the custom object in
the second column (eg. "Syncrip") and the start and stop position relative
to the 3' UTR start in the third and fourth column respecitively (eg. col3
= 20 & col4 = 27, meaning position 1020 to 1027 if the first NT of the
3'UTR is at position 1001)}

\item{species}{character object. Can be "human", "mouse" or "rat" at the
moment}

\item{expressed.miRNAs}{A vector of expressed miRNAs (miRBase IDs)}

\item{co_dist_min}{Minimum number of nucleotides to be located between two
miRNA-seeds / KD-12mers and taken into consideration for the search. By
specifying co_dist_min and co_dist_max, the user can search for certain
types of functional miRNA colocalization. Mutual miRNA blocking is
suggested to happen within a distance of 0-7Nt. Cooperative repression of
two miRNAs is reported to happen within a distance of 8-40Nt, and at bigger
distance, it is suggested to miRNAs act independently an given targets. If
both, co_dist_min and co_dist_max are left blank, enrichment of miRNA-pairs
on the whole transcript will be reported.}

\item{co_dist_max}{Maximum number of nucleotides to be located between two
miRNA-seeds / KD-12mers and taken into consideration for the search. For
further info, see co_dist_min.}

\item{min_pairs_sub}{Minimum number of required miRNA-pairs in the subset of
genes to be displayed in the result table. Default = 10 (bit arbitrary, one
could maybe go down to 5?)}

\item{Possibility}{to use prediced KS-sites for the search. Not yet.}
}
\value{
a data.table of all colocalizations including statistics
}
\description{
Finds pairs of miRNAs that are located on the same gene or transcript within
a distance that the user can specify. Fisher's enrichment test is applied in
order to determine which miRNA-pairs are significantly enriched in the
particular universe of expressed genes.
}
