% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sublists.R
\name{findGO}
\alias{findGO}
\title{findGO}
\usage{
findGO(expr, fixed = FALSE, ontology = "BP|MF|CC")
}
\arguments{
\item{expr}{The expression to query, interpreted as a regular expression if 
`fixed=FALSE`.}

\item{fixed}{Whether to interpret `expr` as a fixed expression, rather than 
a regular expression (default FALSE)}

\item{ontology}{Ontology to search, default all 3 (`BP|MF|CC`)}
}
\value{
A named vector of GO IDs, with Terms as names.
}
\description{
Finds GO categories where the term matches a query expression.
}
