% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{fisher.test.p}
\alias{fisher.test.p}
\title{fisher.test.p}
\usage{
fisher.test.p(a, b, c, d, alternative = c("two.sided", "greater", "less"))
}
\arguments{
\item{a, b, c, f}{Vectors containing the four positions of the 2x2 matrix}

\item{alternative}{greater, less, or 'two.sided' (default)}
}
\value{
A vector of p-values
}
\description{
Fast p-values from multiple Fisher's exact tests
}
