% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{geneBasedTest}
\alias{geneBasedTest}
\title{geneBasedTest}
\usage{
geneBasedTest(features, TS)
}
\arguments{
\item{features}{The set of differentially-expressed features, or features of interest.}

\item{TS}{A data.frame of miRNA targets, with at least the following columns: `family`, `rep.miRNA`, `feature`, `sites`.}
}
\value{
a named vector of p-values.
}
\description{
Applies Fisher's test to the number of miRNA binding sites of each differentially-expressed gene
}
