% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sublists.R
\name{getGOgenes}
\alias{getGOgenes}
\title{getGOgenes}
\usage{
getGOgenes(go_ids, species = "Mm", translate.ids = TRUE)
}
\arguments{
\item{go_ids}{A vector of GO IDs.}

\item{species}{A two-letter code indicating the species, such as `Mm` (default) or `Hs`.}

\item{translate.ids}{Logical; whether to translate GO IDs into terms (default TRUE)}
}
\description{
getGOgenes
}
