% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getHumanMirExp}
\alias{getHumanMirExp}
\title{getHumanMirExp}
\usage{
getHumanMirExp(x = NULL)
}
\arguments{
\item{x}{Desired tissue or celltype}
}
\value{
If `x` is NULL, returns a vector of the possible values. If `x` is 
given and matches a tissue/celltype of the dataset, returns the average miRNA
expression as logCPM, i.e. log(1+counts per million).
}
\description{
Get the human miRNA expression in a given tissue or celltype from the 
`microRNAome` package 
(\href{http://genome.cshlp.org/content/27/10/1769}{McCall et al., 2017}).
}
\examples{
head(getHumanMirExp("thyroid"))
}
