% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getMouseMirExp}
\alias{getMouseMirExp}
\title{getMouseMirExp}
\usage{
getMouseMirExp(x = NULL)
}
\arguments{
\item{x}{Desired tissue or celltype}
}
\value{
If `x` is NULL, returns a vector of the possible values. If `x` is 
given and matches a tissue/celltype of the dataset, returns the average miRNA
expression as logCPM, i.e. log(1+counts per million).
}
\description{
Get the mouse miRNA expression in a given tissue or celltype, based on the
GSE119661 dataset 
(\href{https://doi.org/10.1093/nar/gkaa323}{Kern et al., 2020}) 
supplemented with some celltypes from the GSE30286 dataset 
(\href{https://doi.org/10.1016/j.neuron.2011.11.010}{He et al., 2012}).
}
\examples{
head(getMouseMirExp("Muscle"))
}
