% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichMiR.R
\name{getResults}
\alias{getResults}
\title{getResults}
\usage{
getResults(object, test = NULL, getFeatures = TRUE, flatten = FALSE)
}
\arguments{
\item{object}{An object of class `enrich.results`.}

\item{test}{The name of a test, or a vector of such names. If ommitted (default), a 
merge of all available tests is returned. Note that more detailed results are returned
 when looking at a specific test.}

\item{getFeatures}{Logical; whether to return overlapping features.}

\item{flatten}{Logical; whether to return a normal data.frame instead of a DataFrame.}
}
\value{
a DataFrame or data.frame.
}
\description{
Returns the results table of an enrich.results object.
}
