% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_import.R
\name{importTargetScan}
\alias{importTargetScan}
\title{importTargetScan}
\usage{
importTargetScan(ts.filepath, featureIdentifier = "symbol",
  recapitalizeGenes = FALSE)
}
\arguments{
\item{ts.filepath}{Path to the targetScan conserved targets file.}

\item{featureIdentifier}{The type of feature identifier to use as feature name; either "symbol", "transcript", or "geneid".}

\item{recapitalizeGenes}{Logical; whether to recapitalize genes to the mouse standard (default FALSE)}
}
\value{
a data.frame.
}
\description{
import a targetScan conserved targets file (conserved site and contextScore).
}
