% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{KS}
\alias{KS}
\title{KS}
\usage{
KS(DEA, TS, minSize = 5)
}
\arguments{
\item{DEA}{A data.frame of the results of a differential expression analysis, with features (e.g. genes) as row names and with at least the following columns: `logFC`, `FDR`}

\item{TS}{A data.frame of miRNA targets, with at least the following columns: `family`, `rep.miRNA`, `feature`, `sites`.}

\item{minSize}{The minimum number of targets for a miRNA family to be tested (default 5).}
}
\value{
a data.frame.
}
\description{
miRNA targets enrichment analysis using a Kolmogorov-Smirnov test on the targets' foldchanges. As all alternatives are considered, significance might not always be consistent with differential miRNA activity.
}
