% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{KS2}
\alias{KS2}
\title{KS2}
\usage{
KS2(DEA, TS, minSize = 5)
}
\arguments{
\item{DEA}{A data.frame of the results of a differential expression analysis, with features (e.g. genes) as row names and with at least the following columns: `logFC`, `FDR`}

\item{TS}{A data.frame of miRNA targets, with at least the following columns: `family`, `rep.miRNA`, `feature`, `sites`.}

\item{minSize}{The minimum number of targets for a miRNA family to be tested (default 5).}
}
\value{
a data.frame.
}
\description{
miRNA targets enrichment analysis using a Kolmogorov-Smirnov test on the targets' foldchanges, treating upregulated and downregulated genes separately, and applying a one-sided test to each. Significance here is more likely to incidate an effect consistent with miRNA activity than in a normal KS-test.
}
