% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{lmadd}
\alias{lmadd}
\title{lmadd}
\usage{
lmadd(
  signal,
  sets,
  use.intercept = FALSE,
  calc.threshold = 0.2,
  comb.threshold = 0.05
)
}
\arguments{
\item{signal}{A named numeric vector indicating the signal (e.g. logFC) of
each feature.}

\item{sets}{A data.frame of annotation, with at least the following columns: 
`set`, `feature` and `score`. Alternatively, a sparse numeric matrix, with sets as 
columns and features as rows (dimensions must be named).}

\item{use.intercept}{Logical; whether to use an intercept in the model.}

\item{calc.threshold}{Minimum p-value threshold for the individual fit in 
order to calculate a complementary coefficient.}

\item{comb.threshold}{Minimum p-value threshold for the individual fit in 
order to include as covariate for other coefficients.}
}
\value{
A data.frame.
}
\description{
Complementary linear fits between signal and sets' feature scores
}
