% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{loadTargetScan}
\alias{loadTargetScan}
\title{loadTargetScan}
\usage{
loadTargetScan(
  ts.filepath,
  featureIdentifier = "symbol",
  recapitalizeGenes = FALSE
)
}
\arguments{
\item{ts.filepath}{Path to the targetScan conserved targets file.}

\item{featureIdentifier}{The type of feature identifier to use as feature name; either "symbol", "transcript", or "geneid".}

\item{recapitalizeGenes}{Logical; whether to recapitalize genes to the mouse standard (default FALSE)}
}
\value{
a data.frame.
}
\description{
loads a targetScan conserved targets file (conserved site and contextScore).
}
