% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{matchMirExpr}
\alias{matchMirExpr}
\title{matchMirExpr}
\usage{
matchMirExpr(me, setsNames = NULL, restrict = TRUE)
}
\arguments{
\item{me}{A named vector of miRNA expression, or data.frame with an 
'expression' column}

\item{setsNames}{An optional vector of known sets names, or an actual target
annotation.}

\item{restrict}{Logical indicating whether to restrict `me` to elements 
included in `setNames` (after re-formating)}
}
\value{
A data.frame of miRNA expressions.
}
\description{
Tries to match input miRNA expression with given names, eventually adding the
-3p/-5p.
}
