% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{overlap}
\alias{overlap}
\title{overlap}
\usage{
overlap(signal, sets, alternative = c("greater", "less", "two.sided"))
}
\arguments{
\item{signal}{A named logical vector indicating which features are 
differentially-expressed}

\item{sets}{A data.frame of annotation, with at least the following columns: 
`set`, `feature`. Alternatively, a sparse logical matrix, with sets as 
columns and features as rows (dimensions must be named).}

\item{alternative}{Test alternative (defaults to 'greater' to test for
over-representation)}
}
\value{
a data.frame.
}
\description{
Traditional overlap (Fisher test) between a logical signal (e.g. 
differentially-expressed features) and the elements of sets.
}
