% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KdModels.R
\name{prep12mers}
\alias{prep12mers}
\title{prep12mers}
\usage{
prep12mers(x, mod = NULL, maxSeedMedian = -1.2)
}
\arguments{
\item{x}{A vector of 12-mers, or a data.frame containing at least the columns
"X12mer" and "log_kd"}

\item{mod}{An optional linear model summarizing the kd activity.}

\item{maxSeedMedian}{Max median log_kd for alternative seed inclusion.}
}
\value{
A data.frame
}
\description{
prep12mers
}
