% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{recapitalizeGenes}
\alias{recapitalizeGenes}
\title{recapitalizeGenes}
\usage{
recapitalizeGenes(x, gformat = "mouse")
}
\arguments{
\item{x}{A character vector or an array (in which case the function will be 
applied to row.names)}

\item{gformat}{The gene format, either 'human' (all caps) or 'mouse' (first 
letter capitalized)}
}
\value{
An object of the same type and dimensions as `x`.
}
\description{
A utility function to reformat gene names.
}
