% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{recapitalizeMiRs}
\alias{recapitalizeMiRs}
\title{recapitalizeMiRs}
\usage{
recapitalizeMiRs(x)
}
\arguments{
\item{x}{A character vector or an array (in which case the function will be 
applied to row.names)}
}
\value{
An object of the same type and dimensions as `x`.
}
\description{
A utility function to reformat miRNA names so that there are no capital 
letters except the 'miR-'.
}
