% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{regmir}
\alias{regmir}
\title{regmir}
\usage{
regmir(
  signal,
  sets,
  binary = NULL,
  alpha = 1,
  do.plot = FALSE,
  use.intercept = FALSE,
  keepAll = TRUE
)
}
\arguments{
\item{signal}{A vector of logical or numeric values, with gene symbols as names.}

\item{sets}{A data.frame with at least the following columns: 
'set', 'feature', and (if binary=FALSE) 'score'. Alternatively, a sparse
logical (if binary=TRUE) or numeric matrix with features as rows and sets
as columns.}

\item{binary}{Logical; whether to consider target prediction as binary.}

\item{alpha}{elastic net mixing param (0=ridge, 1=lasso)}

\item{do.plot}{Logical, whether to plot coefficients against lambda}

\item{use.intercept}{Logical, whether to use an intercept in the model.}

\item{keepAll}{Logical, whether to return all families.}
}
\value{
A DataFrame.
}
\description{
miRNA enrichment analysis using regularized regression
}
