% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{siteoverlap}
\alias{siteoverlap}
\title{siteoverlap}
\usage{
siteoverlap(signal, sets)
}
\arguments{
\item{signal}{A named logical vector indicating which features are 
differentially-expressed}

\item{sets}{A data.frame with at least the following columns: 
'set', 'feature', 'sites'.}

\item{alternative}{Test alternative (defaults to 'greater' to test for
over-representation)}
}
\value{
a data.frame.
}
\description{
Applies Fisher's test to the number of binding sites among a set of features 
(vs all other binding sites in that set of features).
This is an old apply-based implementation. For a much faster vectorial 
implementation, see `siteoverlap2`
}
