% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetAnnotation.R
\docType{class}
\name{show,targetAnnotation-method}
\alias{show,targetAnnotation-method}
\alias{targetAnnotationMatrix-class}
\alias{targetAnnotationDF-class}
\alias{show,targetAnnotationDF-method}
\alias{head,targetAnnotationDF-method}
\alias{as.data.frame,targetAnnotationDF-method}
\alias{as.data.frame,targetAnnotationMatrix-method}
\alias{as.matrix,targetAnnotationMatrix-method}
\alias{as.matrix,targetAnnotationDF-method}
\alias{setTargetFeatures,targetAnnotationDF-method}
\alias{setTargetFeatures,targetAnnotationMatrix-method}
\alias{names,targetAnnotationDF-method}
\alias{names,targetAnnotationMatrix-method}
\alias{length,targetAnnotationMatrix-method}
\alias{length,targetAnnotationDF-method}
\alias{[,targetAnnotationDF,ANY,ANY,ANY-method}
\alias{[,targetAnnotationMatrix,ANY,ANY,ANY-method}
\alias{targetAnnotation}
\title{targetAnnotation}
\usage{
\S4method{show}{targetAnnotation}(object)

\S4method{show}{targetAnnotationDF}(object)

\S4method{head}{targetAnnotationDF}(x, n = 6L, ...)

\S4method{as.data.frame}{targetAnnotationDF}(x, name)

\S4method{as.data.frame}{targetAnnotationMatrix}(x, name)

\S4method{as.matrix}{targetAnnotationMatrix}(x, ...)

\S4method{as.matrix}{targetAnnotationDF}(x, ...)

\S4method{setTargetFeatures}{targetAnnotationDF}(x)

\S4method{setTargetFeatures}{targetAnnotationMatrix}(x)

\S4method{names}{targetAnnotationDF}(x)

\S4method{names}{targetAnnotationMatrix}(x)

\S4method{length}{targetAnnotationMatrix}(x)

\S4method{length}{targetAnnotationDF}(x)

\S4method{[}{targetAnnotationDF,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{targetAnnotationMatrix,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

targetAnnotation(
  x,
  content = "score",
  metadata = list(),
  set.properties = NULL,
  feature.synonyms = NULL,
  sets.synonyms = NULL
)
}
\arguments{
\item{x}{A names list of items, or a (sparse) matrix with named dimensions,
or a data.frame with at least the columns `feature` and `set`, or a 
(h5/rds) file containing one of these.}

\item{content}{The content of the matrix (e.g. 'score' or 'sites'); ignored
if `x` is a data.frame or list.}

\item{metadata}{A list of metadata.}

\item{set.properties}{A data.frame of additional set properties.}

\item{feature.synonyms}{A named factor vector of feature synonyms.}

\item{sets.synonyms}{A named factor vector of set names synonyms.}
}
\value{
An object of class `targetAnnotation`; depending on `x`, either a 
`targetAnnotationMatrix` or a `targetAnnotationDF`.
}
\description{
Creates a new targetAnnotation object.
}
