% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findColocalization.R
\name{testColocalization}
\alias{testColocalization}
\title{testColocalization}
\usage{
testColocalization(
  x,
  background = NULL,
  mir_pos,
  sets_pos = NULL,
  mir_pos.properties = NULL,
  th.abs.logFC = 0,
  th.FDR = 0.05,
  minSize = 5,
  maxSize = Inf,
  co_dist_min = 0,
  co_dist_max = NULL,
  force.global.mode = FALSE,
  BP = NULL,
  ...
)
}
\arguments{
\item{x}{The signature in which to look for a signal. This can be either:
\itemize{ \item A data.frame of the results of a differential expression
analysis, with features (e.g. genes) as row names and with at least the
following columns: `logFC`, `FDR`; \item A logical vector of membership to
the geneset of interest, with feature (e.g. genes) as names, in which case
only tests based on a binary signal will be available. \item A vector of
feature names; in this case the \code{background} argument will also be
required, and only tests based on a binary signal will be available. } By
convention, we only test for enrichments in annotated genes.}

\item{background}{A character vector of background; ignored if `x` is not a
character vector.}

\item{mir_pos}{A data.frame (or DataFrame) with positional information of
miRNA Binding Sites as obtained through the findSeedMatches function or
\link{\code{enrichMiR::.getTargetScanPos}}. Should contain at least the following
columns: `set`, `feature`,`start` & `stop`.}

\item{sets_pos}{A data.frame (or DataFrame) / data.table with positional
information of any RBP motif sites as for example obtained through the
""link to RBP_Pos function"" or in custom manner. Should contain at least
the following columns: `set` = RBP or Name, `feature` = Gene or Transcript
Target,`start` & `stop` (positional information, with regards to the
3'UTR).}

\item{mir_pos.properties}{Any further information about the mir_pos object;
this can either be a data.frame (or DataFrame), with row.names
corresponding to names of `sets` (or to alternative names), or a named
vector (e.g. miRNA expression values). If this is given, sets (in the
mir_pos object) not included within this object will be discarded.}

\item{th.abs.logFC}{The minimum absolute log2-foldchange threshold for a
feature to be considered differentially-expressed (default 0). Ignored if
`x` is not a DEA data.frame.}

\item{th.FDR}{The maximum FDR for a feature to be considered differentially-
expressed (default 0.05). Ignored if `x` is not a DEA data.frame.}

\item{minSize}{The minimum size of a set to be tested (default 5).}

\item{maxSize}{The maximum size of a set to be tested (default Inf).}

\item{co_dist_min}{Minimum number of nucleotides to be located between two
miRNA-seeds or a miRNA-seed and RBP-motif. By specifying co_dist_min and
co_dist_max, the user can search for certain types of functional miRNA
colocalization. Mutual miRNA blocking is suggested to happen within a
distance of 0-7Nt. Cooperative repression of two miRNAs is reported to
happen within a distance of 8-40Nt, and at bigger distance, miRNAs are
suggested to act independently on given targets. If co_dist_max = NULL,
enrichment analysis will be performed on the whole transcript.}

\item{co_dist_max}{Maximum number of nucleotides to be located between two
miRNA-seeds or a miRNA-seed and RBP-motif. For further info, see
co_dist_min.}

\item{force.global.mode}{To force a statistical enrichment analysis on the
whole universe of expressed genes, even in case there is for example a
data.frame of the results of a differential expression analysis, specified
as x. By default: FALSE.}

\item{BPPARAM}{could be potentially included, see .findcolocPL at the bottom
\link{BiocParallel} multithreading parameters. Used to multithread
colocalization.}
}
\value{
an enrich.results object.
}
\description{
Creates an enrich.results object and performs enrichment analysis of miRNA
colocalizations.
}
\details{
In detail, finds colocalizations of miRNAs with other miRNAs or user-defined
Partners that are located on the same gene or transcript within a specified
distance. Fisher's enrichment test is applied in order to determine which
miRNA-pairs are significantly enriched in the particular universe of
expressed genes.

In "global" mode, statistical enrichment analysis is performed on the
universe of expressed genes, meaning is there a statistical enrichment of a
certain "pair" with regards to the other pairs that are found in the
universe. The results of this search can be accessed with "Fisher.Pair" in
the enrich.result object.

In "subset" mode, we search for enrichments of pairs in either specified
genes of interest or up- / downregulated genes of a differential expression
analysis. By default, three independent statistics will be performed: 1)
"Fisher.Pair": In order to check whether there is an enrichment of a specific
pair in the subset vs the background. 2) "Fisher.MIR.Co": Check whether the
specified microRNA has more colocolizations with this partner than with any
other. 3) "Fisher.Partner.Co": Check whether the specified Partner has more
colocolizations with this microRNA than with any other.
}
