% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichMiR.R
\name{testEnrichment}
\alias{testEnrichment}
\title{testEnrichment}
\usage{
testEnrichment(x, sets, background = NULL, tests = NULL,
  sets.properties = NULL, th.abs.logFC = 0, th.FDR = 0.05,
  minSize = 5, maxSize = Inf, gsea.maxSize = 1000,
  gsea.permutations = 2000, testOnlyAnnotated = FALSE,
  keepAnnotation = FALSE, BPPARAM = NULL, field = NULL, ...)
}
\arguments{
\item{x}{The signature in which to look for a signal. This can be either:
\itemize{
\item A data.frame of the results of a differential expression analysis, 
with features (e.g. genes) as row names and with at least the following 
columns: `logFC`, `FDR`;
\item A numeric vector of the signal with feature (e.g. genes) as names, in
which case only tests based on a continuous signal will be possible;
\item A logical vector of membership to the geneset of interest, with feature
(e.g. genes) as names, in which case only tests based on a binary signal will
be available.
\item A vector of feature names; in this case the \code{background} argument will
also be required, and only tests based on a binary signal will be available.
}}

\item{sets}{Either a named list (or SimpleList) of features, or a data.frame
(or DataFrame) with at least the columns `set` and `feature` (if the columns 
`score` and `sites` are also present, they will be used by the 
appropriate tests).}

\item{background}{A character vector of background; ignored if `x` is not a
character vector.}

\item{tests}{Character vector of the tests to perform. See 
\link{\code{availableTests}} for the options.}

\item{sets.properties}{Any further information about the sets; this can 
either be a data.frame (or DataFrame), with row.names corresponding to names
of `sets` (or to alternative names), or a named vector (e.g. miRNA expression
 values). If this is given, sets not included within this object will be 
 discarded.}

\item{th.abs.logFC}{The minimum absolute log2-foldchange threshold for a 
feature to be considered differentially-expressed (default 0). Ignored if `x`
is not a DEA data.frame.}

\item{th.FDR}{The maximum FDR for a feature to be considered differentially-
expressed (default 0.05). Ignored if `x` is not a DEA data.frame.}

\item{minSize}{The minimum size of a set to be tested (default 5).}

\item{maxSize}{The maximum size of a set to be tested (default Inf).}

\item{gsea.maxSize}{The maximum number of targets for a miRNA family to be tested using GSEA (default 300).}

\item{gsea.permutations}{The number of permutations for GSEA (default 2000). See `?fgsea` for more information.}

\item{testOnlyAnnotated}{Whether to excluded features that are no set from 
the background (default FALSE).}
}
\value{
an enrich.results object.
}
\description{
Creates an enrich.results object and performs enrichment analysis
}
