% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{wEA}
\alias{wEA}
\title{wEA}
\usage{
wEA(signal, TS, minSize = 5, testOnlyAnnotated = FALSE,
  method = "Wallenius")
}
\arguments{
\item{signal}{A named logical vector indicating which features are differentially-expressed}

\item{TS}{A data.frame of miRNA targets, with at least the following columns: `family`, `rep.miRNA`, `feature`, `sites`.}

\item{minSize}{The minimum number of targets for a miRNA family to be tested (default 5).}

\item{testOnlyAnnotated}{Whether to excluded features that are bound by no miRNA (default FALSE).}

\item{method}{Method for handling then length bias, default "Wallenius". See `?goseq` for more detail.}
}
\value{
a data.frame.
}
\description{
Weighted hypergeometric test adjusted for total number of miRNA bindings sites. (This is done through the `goseq` package)
}
