% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{aREAmir}
\alias{aREAmir}
\title{aREAmir}
\usage{
aREAmir(dea, TS, minSize = 5, pleiotropy = FALSE)
}
\arguments{
\item{TS}{A data.frame of miRNA targets, with at least the following columns: `family`, `rep.miRNA`, `feature`, `sites`.}

\item{minSize}{The minimum number of elements in a set to be tested (default 5).}

\item{DEA}{A data.frame of the results of a differential expression analysis, with features (e.g. genes) as row names and with at least the following columns: `logFC`, `FDR`}

\item{maxSize}{The maximum number of elements in a set to be tested (default 500). If the test takes too long to run, consider setting this.}

\item{fdr.thres}{The FDR threshold below which genes are considered; default 0.2.}

\item{nperm}{The number of permutations, default 2000. The more permutations, the more precise the p-values, in particular for small p-values.}
}
\value{
a data.frame.
}
\description{
analytic Rank-based Enrichment Analysis using a conversion of targetScan 
scores as weights.
}
