% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{CDplot}
\alias{CDplot}
\title{CDplot}
\usage{
CDplot(ll, by = NULL, k = 5, ...)
}
\arguments{
\item{...}{Any other arguments passed to the plot function.}

\item{x}{Either and object of class `enrichMiR`, or a list containing vectors of foldchanges.}

\item{family}{If `x` is an object of class `enrichMiR`, the family to be plotted.}

\item{col}{The vector of colors to be used for the different distributions (defaults to R default colors).}

\item{xlim}{The x axis range to be displayed, default `c(-2,2)`.}

\item{xlab}{The x axis label, defaults to "log2(foldchange)".}

\item{main}{The plot's title, defaults to the family (if given).}

\item{sub}{The plot's subtitle, defaults to the representative miRNA if the `x` is an object of calss `enrichMiR`.}

\item{splitBy}{Split the genes by number of binding sites (`splitBy='sites'`, default), by targetscan score 
(`splitBy='score'`), or simply into target/non-target (`splitBy=NULL`).}
}
\description{
Plots the cumulative foldchange distribution for a given miRNA family or lists of foldchanges.
}
