% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanning.R
\name{characterizeSeedMatches}
\alias{characterizeSeedMatches}
\title{characterizeSeedMatches}
\usage{
characterizeSeedMatches(x, seed, iScore = c(`8mer` = 2.5, `7mer-m8` =
  1.8, `7mer-A1` = 1.5, `6mer` = 1))
}
\arguments{
\item{x}{A factor or character vector of matches, or a data.frame or GRanges
containing a `sequence` column.}

\item{seed}{The seed which sequences of `x` match.}

\item{iScore}{The base scores attributed to different types of matches,
before considering dinucleotides. Should be a numeric vector with the
following names: "8mer", "7mer-m8", "7mer-A1", "6mer"}
}
\value{
A data.frame, or a `GRanges` if `x` is a `GRanges`
}
\description{
characterizeSeedMatches
}
\examples{
characterizeSeedMatches(c("UAAACCACCC","CGAACCACUG"), "AAACCAC")
}
