% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadFunctions.R
\name{.fetch_mirtarbase}
\alias{.fetch_mirtarbase}
\title{importFrom Matrix sparseMatrix miRTarBase}
\usage{
.fetch_mirtarbase(species, returnType = c("dataframe", "matrix"))
}
\description{
importFrom Matrix sparseMatrix miRTarBase
}
