% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichMiR.R
\name{enrichMiR}
\alias{enrichMiR}
\title{enrichMiR}
\usage{
enrichMiR(DEA, TS, miRNA.expression = NULL, families = NULL,
  th.abs.logFC = 0, th.FDR = 0.05, minSize = 5,
  gsea.maxSize = 1000, gsea.permutations = 2000,
  gsea.fdr.thres = 0.2, testOnlyAnnotated = FALSE, tests = NULL,
  cleanNames = FALSE, ...)
}
\arguments{
\item{DEA}{A data.frame of the results of a differential expression analysis, with features
(e.g. genes) as row names and with at least the following columns: `logFC`, `FDR`}

\item{TS}{A data.frame of miRNA targets, with at least the following columns: 
`family`, `rep.miRNA`, `feature`, `sites`.}

\item{miRNA.expression}{A named vector of miRNAs expression values. 
miRNAs not in this vector are assumed to be not expressed in the system, and are not tested.}

\item{families}{A named vector of miRNA families, with individual miRNAs as names. 
If not given, internal data from the package will be used (mouse miRNA families from targetScan).}

\item{th.abs.logFC}{The minimum absolute log2-foldchange threshold for a feature/gene to be 
considered differentially-expressed (default 0).}

\item{th.FDR}{The maximum FDR for a feature/gene to be considered differentially-expressed (default 0.05).}

\item{minSize}{The minimum number of targets for a miRNA family to be tested (default 5).}

\item{gsea.maxSize}{The maximum number of targets for a miRNA family to be tested using GSEA (default 300).}

\item{gsea.permutations}{The number of permutations for GSEA (default 2000). See `?fgsea` for more information.}

\item{gsea.fdr.thres}{The FDR threshold for inclusion in GSEA (default 0.2).}

\item{testOnlyAnnotated}{Whether to excluded features that are bound by no miRNA (default FALSE).}

\item{tests}{Character vector of the tests to perform. Any combination of: 
`overlap`, `wo` (weighted overlap), `siteMir`, `KS`, `KS2`, `MW`, `GSEA`, `modSites`, `modScore`, 
or NULL to perform all tests (default).}

\item{cleanNames}{Logical; whether to remove prefix from all miRNA names (default FALSE).}
}
\value{
an enrichMiR object.
}
\description{
Creates an enrichMiR object and performs miRNA target enrichment analyses
}
