% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichMir.app.server.R,
%   R/enrichMir.app.server.copy.R
\name{enrichMiR.server}
\alias{enrichMiR.server}
\title{enrichMiR.server}
\usage{
enrichMiR.server()

enrichMiR.server()
}
\value{
A shiny server function

A shiny server function
}
\description{
enrichMiR.server

enrichMiR.server
}
