% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{enrichPlot}
\alias{enrichPlot}
\title{enrichPlot}
\usage{
enrichPlot(
  res,
  size.field = "overlap",
  col.field = NULL,
  sig.field = "FDR",
  min.sig.thres = 0.2,
  min.enr.thres = 1,
  label.sig.thres = 0.05,
  label.enr.thres = 2,
  label.field = "family",
  maxLabels = 10,
  opacity = 0.5,
  main = ""
)
}
\arguments{
\item{res}{The results of an enrichment analysis, as obtained by the `enrichMiR.results` function.}

\item{size.field}{The column determining the size of the bubbles (default "overlap").}

\item{col.field}{The column determining the color of the bubbles (default "expression" if present, otherwise "overlap").}

\item{sig.field}{The column to use as significance (default 'FDR').}

\item{min.sig.thres}{The minimum FDR above which to plot (default 0.2).}

\item{min.enr.thres}{The minumum fold-enrichment above which to plot (default 1).}

\item{label.sig.thres}{The FDR threshold above which to plot the names of the enriched families (default 0.05).}

\item{label.enr.thres}{The enrichment threshold above which to plot the names of the enriched families  (default 2).}

\item{label.field}{The field to use as labels (default 'family')}

\item{maxLabels}{The maximum number of labels to plot (default 10).}

\item{opacity}{The opacity of the bubbles, ranging from 0 to 1 (default 0.5).}

\item{main}{The title of the plot.}
}
\description{
Plots the results of an miRNA enrichment analysis
}
