% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichMiR.R
\name{getResults}
\alias{getResults}
\title{getResults}
\usage{
getResults(object, test = NULL, nameCleanFun = function(x) {     x })
}
\arguments{
\item{object}{An object of class `enrichMiR`, as produced by the enrichMiR function.}

\item{test}{The name of a test, or a vector of such names. If ommitted (default), all available tests are returned. Note that more detailed results are returned when looking at a specific test.}

\item{nameCleanFun}{A function to clean the name of miRNAs. By default, nothing is done; you can try `nameCleanFun=.cleanMiRname` to remove prefixes.}
}
\value{
a data.frame.
}
\description{
Returns the results table of an enrich.results object.
}
