% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichMiR.R
\name{getResults}
\alias{getResults}
\title{getResults}
\usage{
getResults(object, test = NULL)
}
\arguments{
\item{object}{An object of class `enrich.results`.}

\item{test}{The name of a test, or a vector of such names. If ommitted (default), all available tests are returned. Note that more detailed results are returned when looking at a specific test.}
}
\value{
a data.frame.
}
\description{
Returns the results table of an enrich.results object.
}
