% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{ks}
\alias{ks}
\title{ks}
\usage{
ks(signal, sets)
}
\arguments{
\item{signal}{A named logical vector indicating the features of interest 
(e.g. which features are differentially-expressed)}

\item{sets}{A data.frame with at least the following columns: 'set', 'feature'.}
}
\value{
a data.frame.
}
\description{
enrichment analysis using a Kolmogorov-Smirnov test on the signal.
}
