% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{michael}
\alias{michael}
\title{michael}
\usage{
michael(signal, TS, minSize = 3, testOnlyAnnotated = FALSE)
}
\arguments{
\item{signal}{A named logical vector indicating which features are differentially-expressed}

\item{TS}{A data.frame of miRNA targets, with at least the following columns: `family`, `rep.miRNA`, `feature`, `sites`.}

\item{minSize}{The minimum number of targets for a miRNA family to be tested (default 5).}

\item{testOnlyAnnotated}{Whether to excluded features that are bound by no miRNA (default FALSE).}

\item{significant}{The set of differentially-expressed features.}

\item{alternative}{Alternative tested, default "greater" overlap.}
}
\value{
a data.frame.
}
\description{
Applies Fisher's test to the number of miRNA binding sites among a set of features (vs all other binding sites in that set of features).
Of note, this application violates the assumptions of Fisher's exact test.
}
