% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{mw}
\alias{mw}
\title{mw}
\usage{
mw(signal, sets)
}
\arguments{
\item{signal}{A named logical vector indicating the features of interest 
(e.g. which features are differentially-expressed)}

\item{sets}{A data.frame with at least the following columns: 
'set', 'feature'.}
}
\value{
a data.frame.
}
\description{
miRNA targets enrichment analysis using a Mann-Whitney / Wilcoxon test on 
the targets' foldchanges.
}
