% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{overlap}
\alias{overlap}
\title{overlap}
\usage{
overlap(signal, sets)
}
\arguments{
\item{signal}{A named logical vector indicating which features are differentially-expressed}

\item{sets}{A data.frame of annotation, with at least the following columns: `set`, `feature`.}
}
\value{
a data.frame.
}
\description{
Traditional overlap (Fisher test) between differentially-expressed features and miRNA targets.
}
