% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{plMod}
\alias{plMod}
\title{plMod}
\usage{
plMod(signature, sets, var = "sites", correctForLength = NULL)
}
\arguments{
\item{signature}{A named numeric vector}

\item{sets}{A data.frame with at least the following columns: 'set', 
'feature', and the variable specified by `var`.}

\item{var}{The independent variable to use, either 'sites' or 'score'.}

\item{correctForLength}{Logical; whether to correct for length (i.e. total
sites). Defaults to TRUE if `var='sites'`, FALSE otherwise.}
}
\value{
a data.frame
}
\description{
Fits linear models between a signal and set variables. This is a flexible but
slow implementations, we recommend \link{ebayes} for a faster one.
}
