% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{regmir}
\alias{regmir}
\title{regmir}
\usage{
regmir(signal, TS, binary = TRUE, alpha = 1, do.plot = FALSE,
  use.intercept = FALSE)
}
\arguments{
\item{signal}{A vector of logical or numeric values, with gene symbols as names.}

\item{TS}{The targetscan target predictions object}

\item{binary}{Logical; whether to consider target prediction as binary (default). The
alternative is to use the score.}

\item{alpha}{elastic net mixing param (0=ridge, 1=lasso)}

\item{do.plot}{Logical, whether to plot coefficients against lambda}

\item{use.intercept}{Logical, whether to use an intercept in the model.}
}
\value{
A DataFrame.
}
\description{
miRNA enrichment analysis using regularized regression
}
